MOD: cdの関数を一部移動
